Refactor processing page title
