--- web/src/collection/components/ProcessingPage/ProcessingPageHeader/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ProcessingPageHeader";
