+++ web/src/common/components/Title/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Title";
